requirements.txt
click==7.1.2
dataclasses==0.8
fastapi==0.63.0
h11==0.12.0
pydantic==1.8.1
PyJWT==2.0.1
pymongo==3.11.3
python-multipart==0.0.5
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
